We updated the way we collect data from market.
Previously we used the data provided by Chipart YouTube's channel.

Now we decided to use the data from Tom's Hardware website for GPU's
performance and a proprietary methodology to get prices based on
Kabum, Pichau and Terabyte prices.

In this commit we modify the text where we explain to the public how
we get this data to reflect our new methodology.